+++ FaceMan.SemanticHub/Generation/ImageGeneration/ImageTaskStatusEnum.cs
@@ -7,7 +7,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace FaceMan.SemanticHub.ModelExtensions.ImageGeneration
+namespace FaceMan.SemanticHub.Generation.ImageGeneration
 {
     [JsonConverter(typeof(TaskStatusConverter))]
     public enum ImageTaskStatusEnum
