+++ FaceMan.SemanticHub/ModelExtensions/OpenAI/Chat/OpenAIChatCompletionService.cs
@@ -1,8 +1,7 @@
 ﻿using DocumentFormat.OpenXml.EMMA;
-
+using FaceMan.SemanticHub.Generation.TextGeneration;
 using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat;
 using FaceMan.SemanticHub.ModelExtensions.QianWen;
-using FaceMan.SemanticHub.ModelExtensions.TextGeneration;
 
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.ChatCompletion;
