+++ FaceMan.SemanticHub/ModelExtensions/TongYi/TongYiClient.cs
@@ -4,9 +4,8 @@
 
 using DocumentFormat.OpenXml.EMMA;
 using DocumentFormat.OpenXml.Spreadsheet;
-
-using FaceMan.SemanticHub.ModelExtensions.ImageGeneration;
-using FaceMan.SemanticHub.ModelExtensions.TextGeneration;
+using FaceMan.SemanticHub.Generation.ImageGeneration;
+using FaceMan.SemanticHub.Generation.TextGeneration;
 using FaceMan.SemanticHub.ModelExtensions.TongYi.Chat;
 using FaceMan.SemanticHub.ModelExtensions.TongYi.Image;
