+++ FaceMan.SemanticHub/ModelExtensions/WenXin/Chat/WenXinChatCompletionService.cs
@@ -3,9 +3,8 @@
 // See the LICENSE file in the project root for more information.
 
 using DocumentFormat.OpenXml.EMMA;
-
+using FaceMan.SemanticHub.Generation.TextGeneration;
 using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat;
-using FaceMan.SemanticHub.ModelExtensions.TextGeneration;
 
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.ChatCompletion;
