+++ FaceMan.SemanticHub/ModelExtensions/ZhiPu/ZhiPuClient.cs
@@ -2,8 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using FaceMan.SemanticHub.Generation.TextGeneration;
 using FaceMan.SemanticHub.ModelExtensions.AzureOpenAI.Chat;
-using FaceMan.SemanticHub.ModelExtensions.TextGeneration;
 using FaceMan.SemanticHub.ModelExtensions.ZhiPu.Chat;
 using System.Net.Http.Json;
 using System.Runtime.CompilerServices;
